requirements.txt
aiogram==3.17.0
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.9.0.post1
filelock==3.16.1
frozenlist==1.5.0
huggingface-hub==0.27.1
idna==3.10
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-huggingface==0.1.2
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
numpy==1.26.4
openai==1.59.6
packaging==24.2
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sentence-transformers==3.3.1
torch==2.5.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.20